requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.3.0
biopython==1.83
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.8.0
filelock==3.18.0
Flask==2.3.3
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
h5py==3.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-huggingface==0.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.0
networkx==3.2.1
numpy==1.23.5
openai==1.97.0
orjson==3.11.0
packaging==23.2
pandas==2.3.1
pillow==10.2.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.13.1
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.47.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.66.4
transformers==4.41.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
